# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Disbursement API",
        "contact": {
            "name": "Disbursement Team",
            "email": "support@email.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/disburse/request": {
            "post": {
                "tags": [
                    "disburse"
                ],
                "operationId": "/disburse/request",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "seller_id": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "remark": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns succes request disbursement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/requestdisburse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/emptyarray"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/requestdisburse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error: Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/emptyarray"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/disburse/status/{transaction_id}": {
            "get": {
                "tags": [
                    "disburse"
                ],
                "operationId": "/disburse/status/transaction_id",
                "parameters": [
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "description": "transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns transaction id status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/transactionstatus"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/emptyarray"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error: Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/emptyarray"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "emptyarray": {},
...
            "requestdisburse": {
                "properties": {
                    "seller_id": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "integer"
                    },
                    "remark": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "transactionstatus": {
                "properties": {
                    "transaction_id": {
                        "type": "integer"
                    },
                    "amount": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    },
                    "receipt": {
                        "type": "string"
                    },
                    "time_served": {
                        "type": "string"
                    },
                    "fee": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}